Rename test entrypoints and tighten popup messaging
- Convert renamed test files to Bun `describe`/`test` entrypoints
- Fix confirm-transaction pending updates to serialize correctly
- Update devcontainer and Bun config for local test setup
