+++ .devcontainer/devcontainer.json
@@ -1,7 +1,10 @@
 {
 	"name": "The Interceptor",
-	"build": {
-		"dockerfile": "Dockerfile"
+	"image": "mcr.microsoft.com/devcontainers/javascript-node:1-20-bullseye",
+	"features": {
+		"ghcr.io/devcontainers-extra/features/bun:1": {
+			"version": "1.3.12"
+		}
 	},
 	"remoteUser": "node"
 }
