+++ bunfig.toml
@@ -1,3 +1,2 @@
 [install]
 exact = true
-ignoreScripts = true
