--- test/tests/backgroundEthAccounts.ts
@@ -1,183 +1,0 @@
-import * as assert from 'assert'
-import { describe, run, runIfRoot, should } from '../micro-should.js'
-
-type Listener = () => void
-type PortMessage = { method?: unknown, result?: unknown, requestId?: unknown }
-
-function installBrowserMock() {
-	const storageState: Record<string, unknown> = {}
-	;(globalThis as typeof globalThis & { browser: typeof globalThis.browser }).browser = {
-		runtime: {
-			lastError: null,
-			async sendMessage() {
-				return undefined
-			},
-			getManifest: () => ({ manifest_version: 3 }),
-			onMessage: { addListener: (_listener: Listener) => undefined, removeListener: (_listener: Listener) => undefined },
-			onConnect: { addListener: (_listener: Listener) => undefined, removeListener: (_listener: Listener) => undefined },
-		},
-		storage: {
-			local: {
-				async get(keys?: string | string[] | Record<string, unknown> | null) {
-					if (keys === undefined || keys === null) return { ...storageState }
-					if (Array.isArray(keys)) return Object.fromEntries(keys.filter((key) => key in storageState).map((key) => [key, storageState[key]]))
-					if (typeof keys === 'string') return keys in storageState ? { [keys]: storageState[keys] } : {}
-					return Object.fromEntries(Object.entries(keys).map(([key, defaultValue]) => [key, key in storageState ? storageState[key] : defaultValue]))
-				},
-				async set(items: Record<string, unknown>) {
-					Object.assign(storageState, items)
-				},
-				async remove(keys: string | string[]) {
-					for (const key of Array.isArray(keys) ? keys : [keys]) delete storageState[key]
-				},
-			},
-		},
-		tabs: {
-			async query() { return [] },
-			async get() { return undefined },
-			async update() { return undefined },
-			onUpdated: { addListener: (_listener: Listener) => undefined, removeListener: (_listener: Listener) => undefined },
-			onRemoved: { addListener: (_listener: Listener) => undefined, removeListener: (_listener: Listener) => undefined },
-		},
-		windows: {
-			async get() { return undefined },
-			async update() { return undefined },
-		},
-		action: {
-			async setIcon() { return undefined },
-			async setTitle() { return undefined },
-			async setBadgeText() { return undefined },
-			async setBadgeBackgroundColor() { return undefined },
-		},
-		browserAction: {
-			async setIcon() { return undefined },
-			async setTitle() { return undefined },
-			async setBadgeText() { return undefined },
-			async setBadgeBackgroundColor() { return undefined },
-		},
-	} as unknown as typeof globalThis.browser
-	;(globalThis as typeof globalThis & { chrome: { runtime: { id: string } } }).chrome = { runtime: { id: 'test-extension' } }
-	;(globalThis as typeof globalThis & { location: Location }).location = { origin: '' } as unknown as Location
-}
-
-async function loadModules() {
-	return {
-		...await import('../../app/ts/background/background.js'),
-		...await import('../../app/ts/background/backgroundUtils.js'),
-		...await import('../../app/ts/background/settings.js'),
-		...await import('../../app/ts/background/storageVariables.js'),
-	}
-}
-
-function createPort(tabId: number, onPostMessage?: (message: PortMessage) => void) {
-	const messages: PortMessage[] = []
-	const port = {
-		name: '0x0',
-		sender: { tab: { id: tabId } },
-		postMessage(message: unknown) {
-			const typedMessage = message as PortMessage
-			messages.push(typedMessage)
-			onPostMessage?.(typedMessage)
-		},
-	} as unknown as browser.runtime.Port
-	return { port, messages }
-}
-
-function createSimulatorWithGetBlockCounter(getBlockCalls: { count: number }) {
-	return {
-		ethereum: {
-			isBlockPolling: () => true,
-			setBlockPolling: (_value: boolean) => undefined,
-			async getBlock() {
-				getBlockCalls.count += 1
-				return null
-			},
-		},
-	} as const
-}
-
-export async function main() {
-	describe('background eth_accounts', () => {
-		should('skip simulation state refresh for eth_accounts in simulation mode', async () => {
-			installBrowserMock()
-			const { handleInterceptedRequest, websiteSocketToString, changeSimulationMode, setUseSignersAddressAsActiveAddress, updateWebsiteAccess } = await loadModules()
-			const websiteOrigin = 'https://example.test'
-			const website = { websiteOrigin, icon: undefined, title: undefined }
-			const account = 0x1111111111111111111111111111111111111111n
-			await changeSimulationMode({ simulationMode: true, activeSimulationAddress: account, activeSigningAddress: undefined })
-			await setUseSignersAddressAsActiveAddress(false)
-			await updateWebsiteAccess(() => [{ website, access: true, addressAccess: undefined }])
-
-			const socket = { tabId: 1, connectionName: 0n }
-			const { port, messages } = createPort(socket.tabId)
-			const connectionKey = websiteSocketToString(socket)
-			const websiteTabConnections = new Map([[socket.tabId, { connections: {
-				[connectionKey]: { port, socket, websiteOrigin, approved: true, wantsToConnect: true },
-			} }]])
-			const getBlockCalls = { count: 0 }
-			const request = {
-				interceptorRequest: true,
-				usingInterceptorWithoutSigner: false,
-				uniqueRequestIdentifier: { requestId: 1, requestSocket: socket },
-				method: 'eth_accounts',
-			}
-
-			await handleInterceptedRequest(port, websiteOrigin, website, createSimulatorWithGetBlockCounter(getBlockCalls) as unknown as Parameters<typeof handleInterceptedRequest>[3], socket, request, websiteTabConnections)
-
-			assert.equal(getBlockCalls.count, 0)
-			assert.equal(messages.some((message) => message.method === 'request_signer_to_eth_accounts'), false)
-			const ethAccountsReplies = messages.filter((message) => message.method === 'eth_accounts' && message.requestId === 1)
-			assert.deepEqual(ethAccountsReplies.at(-1)?.result, ['0x1111111111111111111111111111111111111111'])
-		})
-
-		should('refresh signer accounts for approved eth_accounts requests when the tab cache is empty', async () => {
-			installBrowserMock()
-			const {
-				handleInterceptedRequest,
-				websiteSocketToString,
-				sendInternalWindowMessage,
-				changeSimulationMode,
-				setUseSignersAddressAsActiveAddress,
-				updateWebsiteAccess,
-				updateTabState,
-			} = await loadModules()
-			const websiteOrigin = 'https://example.test'
-			const website = { websiteOrigin, icon: undefined, title: undefined }
-			const account = 0x2222222222222222222222222222222222222222n
-			await changeSimulationMode({ simulationMode: false, activeSimulationAddress: undefined, activeSigningAddress: undefined })
-			await setUseSignersAddressAsActiveAddress(false)
-			await updateWebsiteAccess(() => [{ website, access: true, addressAccess: undefined }])
-
-			const socket = { tabId: 1, connectionName: 0n }
-			const { port, messages } = createPort(socket.tabId, (message) => {
-				if (message.method !== 'request_signer_to_eth_accounts') return
-				void (async () => {
-					await updateTabState(socket.tabId, (previousState) => ({ ...previousState, signerAccounts: [account], activeSigningAddress: account }))
-					sendInternalWindowMessage({ method: 'window_signer_accounts_changed', data: { socket } })
-				})()
-			})
-			const connectionKey = websiteSocketToString(socket)
-			const websiteTabConnections = new Map([[socket.tabId, { connections: {
-				[connectionKey]: { port, socket, websiteOrigin, approved: true, wantsToConnect: true },
-			} }]])
-			const request = {
-				interceptorRequest: true,
-				usingInterceptorWithoutSigner: false,
-				uniqueRequestIdentifier: { requestId: 7, requestSocket: socket },
-				method: 'eth_accounts',
-			}
-
-			await handleInterceptedRequest(port, websiteOrigin, website, createSimulatorWithGetBlockCounter({ count: 0 }) as unknown as Parameters<typeof handleInterceptedRequest>[3], socket, request, websiteTabConnections)
-
-			assert.equal(messages.filter((message) => message.method === 'request_signer_to_eth_accounts').length, 1)
-			assert.equal(messages.some((message) => message.method === 'request_signer_to_eth_requestAccounts'), false)
-			const ethAccountsReplies = messages.filter((message) => message.method === 'eth_accounts' && message.requestId === 7)
-			assert.deepEqual(ethAccountsReplies.at(-1)?.result, ['0x2222222222222222222222222222222222222222'])
-		})
-	})
-}
-
-await runIfRoot(async () => {
-	await main()
-	await run()
-}, import.meta)
