+++ test/tests/nethermindComparison.test.ts
@@ -61,100 +61,96 @@ class MockEthereumJSONRpcRequestHandler {
 	}
 }
 
-export async function main() {
-	const blockNumber = 8443561n
-	const rpcNetwork = {
-		name: 'Goerli',
-		chainId: 5n,
-		httpsRpc: 'https://rpc.dark.florist/flipcardtrustone',
-		currencyName: 'Goerli Testnet ETH',
-		currencyTicker: 'GÖETH',
-		primary: true,
-		minimized: true,
-		weth: 0xb4fbf271143f4fbf7b91a5ded31805e42b2208d6n,
-	}
-
-	const ethereum = new EthereumClientService(new MockEthereumJSONRpcRequestHandler(), async () => {}, async () => {}, rpcNetwork)
-
-	const simulationState: SimulationState = {
-		blockNumber: blockNumber,
-		blockTimestamp: new Date(0),
-		rpcNetwork: rpcNetwork,
-		baseFeePerGas: 0n,
-		simulationConductedTimestamp: new Date(0),
-		success: true,
-		simulatedBlocks: [],
-		simulationStateInput: [],
-	}
-
-	const exampleTransaction = {
-		type: '1559',
-		from: 0xd8da6bf26964af9d7eed9e03e53415d37aa96045n,
-		nonce: 0n,
-		maxFeePerGas: 1n,
-		maxPriorityFeePerGas: 1n,
-		gas: 21000n,
-		to: 0xda9dfa130df4de4673b89022ee50ff26f6ea73cfn,
-		value: 10n,
-		input: new Uint8Array(0),
-		chainId: 5n,
-	} as const
-
-	describe('Nethermind testing', () => {
-
-		test('getBlock with true aligns with Nethermind', async () => {
-			const block = await getSimulatedBlock(ethereum, undefined, simulationState, blockNumber, true)
-			if (block === null) throw new Error('Block was null')
-			const serialized = GetBlockReturn.serialize(block)
-			const expected = parseRequest(eth_getBlockByNumber_goerli_8443561_true)
-			assertIsObject(expected)
-			assert.equal(JSON.stringify(serialized, Object.keys(block).sort()), JSON.stringify(expected, Object.keys(expected).sort()))
-		})
-
-		test('getBlock with false aligns with Nethermind', async () => {
-			const block = await getSimulatedBlock(ethereum, undefined, simulationState, blockNumber, false)
-			if (block === null) throw new Error('Block was null')
-			const serialized = GetBlockReturn.serialize(block)
-			const expected = parseRequest(eth_getBlockByNumber_goerli_8443561_false)
-			assertIsObject(expected)
-			assert.equal(JSON.stringify(serialized, Object.keys(block).sort()), JSON.stringify(expected, Object.keys(expected).sort()))
-		})
-
-		test('getBlock rejects non-spec 7702 authorization parity values', async () => {
-			await assert.rejects(async () => await ethereum.getBlock(undefined, 8443562n, true))
-		})
-
-		test('adding transaction and getting the next block should include all the same fields as Nethermind', async () => {
-			const block = await getSimulatedBlock(ethereum, undefined, simulationState, blockNumber, true)
-			if (block === null) throw new Error('Block was null')
-			const newState = await appendTransactionToInputAndSimulate(ethereum, undefined, simulationState.simulationStateInput, [{
-				signedTransaction: mockSignTransaction(exampleTransaction),
-				website: { websiteOrigin: 'test', icon: undefined, title: undefined },
-				created: new Date(),
-				originalRequestParameters: { method: 'eth_sendTransaction', params: [{}]},
-				transactionIdentifier: 1n,
-			}])
-			const nextBlock = await getSimulatedBlock(ethereum, undefined, newState, blockNumber + 2n, true)
-			if (nextBlock === null) throw new Error('Block was null')
-			assert.equal(JSON.stringify(Object.keys(nextBlock).sort()), JSON.stringify(Object.keys(block).sort()))
-
-			const expected = parseRequest(eth_getBlockByNumber_goerli_8443561_true)
-			assertIsObject(expected)
-			const requiredFields = Object.keys(expected).sort()
-			assert.equal(JSON.stringify(Object.keys(nextBlock).sort()), JSON.stringify(requiredFields))
-		})
-
-		test('get transaction by hash aligns with Nethermind', async () => {
-			const transaction = await getSimulatedTransactionByHash(ethereum, undefined, simulationState, 0xe10c2a85168046080235fff99e2e14ef1e90c8cf5e9d675f2ca214e49e555e0fn)
-			if (transaction === null) throw new Error('Transaction is not found')
-
-			const serialized = serialize(EthereumSignedTransactionWithBlockData, transaction)
-			const expected = parseRequest(eth_transactionByhash0xe10c2a85168046080235fff99e2e14ef1e90c8cf5e9d675f2ca214e49e555e0f)
-			assertIsObject(expected)
-			assertIsObject(serialized)
-			assert.equal(JSON.stringify(serialized, Object.keys(serialized).sort()), JSON.stringify(expected, Object.keys(expected).sort()))
-		})
-	})
+const blockNumber = 8443561n
+const rpcNetwork = {
+	name: 'Goerli',
+	chainId: 5n,
+	httpsRpc: 'https://rpc.dark.florist/flipcardtrustone',
+	currencyName: 'Goerli Testnet ETH',
+	currencyTicker: 'GÖETH',
+	primary: true,
+	minimized: true,
+	weth: 0xb4fbf271143f4fbf7b91a5ded31805e42b2208d6n,
 }
 
-await main()
+const ethereum = new EthereumClientService(new MockEthereumJSONRpcRequestHandler(), async () => {}, async () => {}, rpcNetwork)
+
+const simulationState: SimulationState = {
+	blockNumber: blockNumber,
+	blockTimestamp: new Date(0),
+	rpcNetwork: rpcNetwork,
+	baseFeePerGas: 0n,
+	simulationConductedTimestamp: new Date(0),
+	success: true,
+	simulatedBlocks: [],
+	simulationStateInput: [],
+}
+
+const exampleTransaction = {
+	type: '1559',
+	from: 0xd8da6bf26964af9d7eed9e03e53415d37aa96045n,
+	nonce: 0n,
+	maxFeePerGas: 1n,
+	maxPriorityFeePerGas: 1n,
+	gas: 21000n,
+	to: 0xda9dfa130df4de4673b89022ee50ff26f6ea73cfn,
+	value: 10n,
+	input: new Uint8Array(0),
+	chainId: 5n,
+} as const
+
+describe('Nethermind testing', () => {
+
+	test('getBlock with true aligns with Nethermind', async () => {
+		const block = await getSimulatedBlock(ethereum, undefined, simulationState, blockNumber, true)
+		if (block === null) throw new Error('Block was null')
+		const serialized = GetBlockReturn.serialize(block)
+		const expected = parseRequest(eth_getBlockByNumber_goerli_8443561_true)
+		assertIsObject(expected)
+		assert.equal(JSON.stringify(serialized, Object.keys(block).sort()), JSON.stringify(expected, Object.keys(expected).sort()))
+	})
+
+	test('getBlock with false aligns with Nethermind', async () => {
+		const block = await getSimulatedBlock(ethereum, undefined, simulationState, blockNumber, false)
+		if (block === null) throw new Error('Block was null')
+		const serialized = GetBlockReturn.serialize(block)
+		const expected = parseRequest(eth_getBlockByNumber_goerli_8443561_false)
+		assertIsObject(expected)
+		assert.equal(JSON.stringify(serialized, Object.keys(block).sort()), JSON.stringify(expected, Object.keys(expected).sort()))
+	})
+
+	test('getBlock rejects non-spec 7702 authorization parity values', async () => {
+		await assert.rejects(async () => await ethereum.getBlock(undefined, 8443562n, true))
+	})
+
+	test('adding transaction and getting the next block should include all the same fields as Nethermind', async () => {
+		const block = await getSimulatedBlock(ethereum, undefined, simulationState, blockNumber, true)
+		if (block === null) throw new Error('Block was null')
+		const newState = await appendTransactionToInputAndSimulate(ethereum, undefined, simulationState.simulationStateInput, [{
+			signedTransaction: mockSignTransaction(exampleTransaction),
+			website: { websiteOrigin: 'test', icon: undefined, title: undefined },
+			created: new Date(),
+			originalRequestParameters: { method: 'eth_sendTransaction', params: [{}]},
+			transactionIdentifier: 1n,
+		}])
+		const nextBlock = await getSimulatedBlock(ethereum, undefined, newState, blockNumber + 2n, true)
+		if (nextBlock === null) throw new Error('Block was null')
+		assert.equal(JSON.stringify(Object.keys(nextBlock).sort()), JSON.stringify(Object.keys(block).sort()))
+
+		const expected = parseRequest(eth_getBlockByNumber_goerli_8443561_true)
+		assertIsObject(expected)
+		const requiredFields = Object.keys(expected).sort()
+		assert.equal(JSON.stringify(Object.keys(nextBlock).sort()), JSON.stringify(requiredFields))
+	})
+
+	test('get transaction by hash aligns with Nethermind', async () => {
+		const transaction = await getSimulatedTransactionByHash(ethereum, undefined, simulationState, 0xe10c2a85168046080235fff99e2e14ef1e90c8cf5e9d675f2ca214e49e555e0fn)
+		if (transaction === null) throw new Error('Transaction is not found')
+
+		const serialized = serialize(EthereumSignedTransactionWithBlockData, transaction)
+		const expected = parseRequest(eth_transactionByhash0xe10c2a85168046080235fff99e2e14ef1e90c8cf5e9d675f2ca214e49e555e0f)
+		assertIsObject(expected)
+		assertIsObject(serialized)
+		assert.equal(JSON.stringify(serialized, Object.keys(serialized).sort()), JSON.stringify(expected, Object.keys(expected).sort()))
+	})
+})
